Introduce standard_java_library() macro.
Summary:Introduce a macro in `DEFS` so we can remove a bunch of boilerplate
in most of our `BUCK` files in the `srcs/` directory.

Applied the macro to all `BUCK` files that:

* Define only one `java_library()` rule.
* No other rule in the build file has the name of the rule. (`src/com/facebook/buck/tools/dxanalysis/BUCK` was an issue here.)
* Use `autodeps = True`.

It turns out that this applied to 48/100 `BUCK` files in the `src/` directory.
Hopefully this sufficiently lowers the barrier to refactoring such that we'll
break up some of the `BUCK` files that don't use `standard_java_library()`
so it can be used in more places.

Had to run `buck autodeps` after rewriting all of the `BUCK` files,
as some of the rules were renamed as a result of introducing the macro.

Test Plan: CI

Reviewed By: yiding

fb-gh-sync-id: aba92cc
fbshipit-source-id: aba92cc