TargetNode's equals & hashcode go back to default
Summary:TargetNode equals() and hashcode() don't take into account the contents of the Arg. The reason the current implementation of these method worked up till now is because the DaemonicParserState would only allow a single instance of a TargetNode to be associated with a given BuildTarget, therefore the BuildTarget was the unique identity of the TargetNode.

The current implementations of hashCode and equals make it impossible to use the TargetNode as a 'key' for caching the ActionGraph, as two TargetNodes with different Args will be equal as long as their BuildTargets match. Using the default implementations derived from Object solves that issue. The downside of this approach is that we are overly-pessimistic: TargetNodes with equal Args will still appear as different, but this approach will serve for now.

Test Plan: CI

Reviewed By: marcinkosiba

fb-gh-sync-id: b934256
fbshipit-source-id: b934256