Set HGPLAIN=1 when running Mercurial commands.
Summary:Specifying `$HGPLAIN` is something that we found to be critical on Nuclide
where we shell out to `hg` for all sorts of things as part of the Hg integration.
Without it, you have no idea what an `hg` subcommand can do because of local aliases
the user may have defined.

Test Plan:`HgCmdLineInterfaceIntegrationTest#testTimestamp()` has been failing on my machine for months.
I've been ignoring it, but I finally decided to investigate and this fixed it.
I have quite a few aliases defined in my `~/.hgrc` file.

Reviewed By: marcinkosiba

fb-gh-sync-id: ae27af2
fbshipit-source-id: ae27af2