Don't allow merging of non-coalescant folders
Summary:It's currently possible for the simplifier to merge; a/b/c, a/b/d, and a/b/d/res, into a/b/ and a/b/d/res which causes intelliJ to complain about a/b/d/res being within the a/b source root.

This diff stops a/b/c/d being collapsed, and so avoids the overlapping source roots.

Test Plan:- Local testing
- Sandcastle builds

Reviewed By: marcinkosiba

fb-gh-sync-id: 325f0d5
fbshipit-source-id: 325f0d5