Allow to specify runnable path for Xcode workspace
Summary:This diff allows to specify runnable path for the application that debugger should attach to.
Also it allows to specify the launch style, automatic or wait, to match parity with Xcode.

Test Plan:CI, added tests

For `xcode_workspace_config` add the arguments

  explicit_runnable_path = '/Applications/Xcode.app',
  launch_style = 'wait',

Generate project and make sure resulting workspace scheme's application is set to Xcode and Wait for process to be executed setting is selected.

Reviewed By: Coneko

fb-gh-sync-id: 3aae9e4
fbshipit-source-id: 3aae9e4