Fix rule keys for native relinker
Summary:Stringifying `CxxPlatform` in the rule key meant that it called
`PreprocessorProvider.toString()` (and other `toString` methods).
The providers however do not implement `toString`, so the rule key
ended up containing the address of the provider objects. This meant
that the rule key would change every time the Buck daemon was
restarted. Fix it by only including `objdump` (a `VersionedTool`)
in the rule key.

Test Plan:Run `buck targets --show-rulekey` to confirm that the rule key
does not change after `buck kill`.

fb-gh-sync-id: c43368f
fbshipit-source-id: c43368f