Fix Python test filtering
Summary:Test filtering was broken because it was geting the list of tests
from `ShellStep`, but `ShellStep` does not make its stdout machine
readable and depending on the configuration it might prepend each
line with ansi color escape codes, breaking parsing of the list of
tests. To fix that, directly use `ProcessExecutor` (and pass it
`EXPECTING_STD_OUT`) instead of a `ShellStep`.

Test Plan: `buck test -f '#test_global_help' //programs:test`

Reviewed By: bhamiltoncx

fb-gh-sync-id: f4cb79c
fbshipit-source-id: f4cb79c