Watchman for ActionGraphCache
Summary:ActionGraphCache has to be invalidated everytime we have a file modification like add, delete or overflow. The cache will not be invalidated if the change is a content modification of a file. The reasoning is that content modifications do not alter the graphs.

This check at the ActionGraphCache is an extra level of checking. The normal scenario is that the DaemonicParser will receive the same events and gets invalidated before reaching the ActionGraph. Still, better to be sure in case the
flow changes in the future.

Test Plan: Added a unit-test that checks that cache gets invalidated on every WatchEvent except content modification of a file. It also check that the counters count those misses and the one hit in the end.

Reviewed By: marcinkosiba

fb-gh-sync-id: 8a0882f
fbshipit-source-id: 8a0882f