sh_binary: don't chmod tmp dir contents to read-only
Summary:The generated shell binary helper script would chmod the the temp dir
it setup with symlinks to read-only, then chmod it back to writable and
clean it via a trap handler.  This meant that if the script was failed
in some way, it was possible it left it's contents around as read-only,
making things like `RmStep` fail.

This diff just removes the chmod'ing altogether, leaving the tmp dir
writable.  Since it only contained symlinks and since scripts
could also chmod it to writable anyway, it's not clear that the original
chmod added much.

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: e85cf37
fbshipit-source-id: e85cf37