Use precompiled headers when prefix header is set.
Summary:This adds support for precompiled headers on toolchains that supports it
(currently clang). When compiling in combined mode, and prefix header exists, a
pch file is generated from it for each language and each combination of
compiler flags.

The pch file is not cached due to it being difficult to sanitize and normalize,
and because they are very fast to generate.

Test Plan: added unit and integration tests

Reviewed By: Coneko

fb-gh-sync-id: 6a93a36
fbshipit-source-id: 6a93a36