CachingBuildEngine: only log hash of cacheable build rules
Summary:We log the hash of a rules outputs to help detect non-deterministic
rules.  However this can add some significant overhead.  To help mitigate
this, don't grab the output hash of uncacheable rules, as they are likely
not cached due to the sheer size which would be costly to hash or builtin
non-determinism in the rule which somewhat defeats the purpose of logging
the hash.

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: cb57ea9
fbshipit-source-id: cb57ea9