Make completion work on zsh
Summary:Little is needed to make the tab completion work on ZSH as well. This would be useful for me (at least).

I've added reviewers according to file history - you are welcome to move to someone else if there is someone more appropriate.
Thanks!

Test Plan:On both bash and zsh do:
1) source ~/devtools/buck/scripts/buck-completion.bash
2) buck <tab>
   List of verbs should appear
3) buck b<tab>
   Should complete to build
4) buck build <tab>
   Many targets should appear
5) buck build sp<tab>
   Targets starting with sp should appear
6) buck build spa<tab>
   Should complete to spaceship.

Reviewed By: k21

fb-gh-sync-id: 4f67e19
fbshipit-source-id: 4f67e19