Learn to read source.properties for new NDK
Summary:Android r11 and later changed the file and version representation. In previous versions `RELEASE.TXT` hold the version, in r11+ it is `source.properties`. Now, the method acknowledges this and try to read from the new model, if the file isn't available it falls back to the older `RELEASE.TXT`.

The code gives priority to `source.properties`. In case you have both on the NDK dir the `source.properties` will be read.

Test Plan:Added/extended a few unit tests. Cases:
- You read a config with the new way.
- You pick between an old NDK and a new one correctly.
- You find the latest properly.
- Pick the correct new one even if it not an exact match.

Reviewed By: Coneko

fb-gh-sync-id: 5d98442
fbshipit-source-id: 5d98442