Record process time out directly from the waiter thread in ProcessExecutor
Summary:Instead of interrupting the waiter thread, waiting for it to exit, and then
look if process is still running to determine whether the process as timed out,
record the timeout when the waiter thread is interrupted by the calling thread.

This essentially cuts the time it takes to do thread join, which may improve timeout
detection when the system is under contention during test runs.

Test Plan: CI

Reviewed By: bolinfest

fb-gh-sync-id: c55b391
fbshipit-source-id: c55b391