Reduce the amount of graph traversal when hashing targets
Summary:The existing hashing algorithm does a lot of excessive hashing.

- At first stage it walk the graph and does bottom-up hashing of all targets (using depth-first post-order traversal); this hashing accepts target's information (input files and such) and target's dependencies hashes.
- Later for every input target it builds dependency closure (which includes all targets from target's dependencies), adds test targets if needed and hashes the result set using hashes computed in the first step.

This diff makes the second stage optional. It is performed only if `--detect-test-changes` is passed in arguments.

Also, the second stage logic is now similar to the first stage. Instead of creating a dependency closure for every input target, it does depth-first post-order traversal (yes, second time) and hashes targets using target's dependencies and tests. This is done in order to make target's hash depends on test targets of its dependencies.

Test Plan:That's a bit difficult test right now because `TargetsCommand` doesn't allow hashing all targets in the graph. Passing a list of targets on the other hand makes parsing a bottleneck.

I tested this change using the last change in the stack; this change adds support for hashing when there are no arguments in the command line. It makes run time from infinite (more than 1 hour) down to 5 minutes.

Tested with this command:

  time NO_BUCKD=1 buck targets --detect-test-changes --show-target-hash

Reviewed By: Coneko

fb-gh-sync-id: fee41af
fbshipit-source-id: fee41af