ClassUsageFileWriter
Summary:
Instead of passing an `Optional<Path>` and having `Jsr199Javac` key off of that to write
the class usage file, pass a `ClassUsageFileWriter` (which may or may not be a no-op
implementation). Setting it up this way allows us to have `DefaultJavaLibrary` own the
`ClassUsageTracker`, which will be helpful when it comes time to implement
`SupportsDependencyFileRuleKey.getInputsAfterBuildingLocally`.

Test Plan: Existing unit tests should cover it

Reviewed By: marcinkosiba

fb-gh-sync-id: dc1e734
fbshipit-source-id: dc1e734