Early-out right before executing steps.
Summary:It can take multiple seconds between scheduling the execution of
a rule and submitting the execution of the steps to the executor.
In such a case the 'fistFailure' early-out path was not being
taken causing the build to keep going way past the first failure.

Test Plan: manual testing

Reviewed By: k21

fb-gh-sync-id: 889a545
fbshipit-source-id: 889a545