Have ReactNativeBundle return bundle path for getPathToOutput
Summary: The ##getPathToOutput## for ##ReactNativeBundle## initially returned ##null## because there's slight ambiguity given that the rule actually generates multiple outputs in sibling directories (the bundle file, the assets, and the source map). Without a return value, ##buck targets --show-output## doesn't print anything. This diff makes it so that ##getPathToOutput## returns the path to the JS bundle since that is arguably the most important output of the rule.

Test Plan:Added test case:
  buck test //test/com/facebook/buck/js

Reviewed By: Coneko

fb-gh-sync-id: 40fec64
fbshipit-source-id: 40fec64