Prepare DepFile serialization for Java Class ABI dep files
Summary:
The dependency file serialization code currently assumes that each entry
in the dep file is a path to a file on disk. As we add class-level Java ABI
support, this assumption will no longer hold -- now each entry may be
a path to a member of an archive.

This diff introduces a `DependencyFileEntry`, an immutable class that exists
pretty much only so that we can easily use Jackson to write this information
to metadata.

Test Plan:
This diff adds no significant new functionality; existing tests should be
sufficient to ensure no regressions.

Reviewed By: marcinkosiba

fb-gh-sync-id: 459e3c4
fbshipit-source-id: 459e3c4