Fix support for using .whl files in prebuilt_python_library
Using wheel files with prebuilt python libraries results in the
following traceback:

```
Traceback (most recent call last):
  File ".bootstrap/_pex/pex.py", line 314, in execute
  File ".bootstrap/_pex/pex.py", line 78, in _activate
  File ".bootstrap/_pex/environment.py", line 132, in activate
  File ".bootstrap/_pex/environment.py", line 184, in _activate
  File ".bootstrap/_pex/environment.py", line 147, in _resolve
  File ".bootstrap/pkg_resources/__init__.py", line 854, in resolve
  File ".bootstrap/pkg_resources/__init__.py", line 2617, in requires
  File ".bootstrap/pkg_resources/__init__.py", line 2818, in _dep_map
  File ".bootstrap/pkg_resources/__init__.py", line 2835, in _compute_dependencies
ImportError: No module named _markerlib
```

This is because we did not also vendor _markerlib from 18.5 of
setuptools.

This commit pulls in _markerlib and adds it to the pex at the same time
that pkg_resources is added.

Fixes #719