CachingBuildEngine: cache rule output hashes in artifact metadata
Summary:
We hash the contents of rule outputs for use in input-based rule
keys.  When the outputs for rules are large, this means we sometimes
spend a lot of time calculating the input-based rule key, as we're
hashing lots of outputs of other rules in the serial rule key calculation
step a dependent rule.  This was partially mitigated by some code we
had to log the output hash of a rule after it had been built locally,
as it distributed the hashing workload across several threads.
However, this optimization didn't happen when fetching artifacts from
the cache, and so builds with lots of cache hits would then get hit
by the large input-based rule key calculation step.

This diff stores the hashes of rule outputs into the cached artifact
metadata, so that they can be used to seed the local file hash cache
on a cache fetch.  This means that we should now always avoid long
input-based rule key calculations (as rules are either build locally,
which ditributes the hashing operation, or can benefit from the
cached output hash).

Test Plan: added unittest

Reviewed By: marcinkosiba

fb-gh-sync-id: 47f4e18
fbshipit-source-id: 47f4e18