Allow --show-target-hash without targets in arguments
Summary:
This change extends `TargetsCommand` to allow running it without arguments when `--show-target-hash` is specified.

Instead of throwing an exception when there are no arguments, the graph nodes are filtered according to specified filters (references files, target type, etc.) and are passed as input targets.

Test Plan:
`buck targets --show-target-hash`

Run twice with some changed between the runs. Make sure the only difference is in targets that depend on changed file.

Reviewed By: k21

fb-gh-sync-id: 9c295b4
fbshipit-source-id: 9c295b4