haskell: make empty archive link rule fix more robust
Summary:
We use a `-optl` to have the haskell linker pass all the linker flags
down into the system linker.  However, since the haskell linker needs
at least one normal argument to run, we worked around this by generating
a dummy empty library that it'd consume but which would eventually get
thrown away at link time (as it wouldn't provide any used symbols).

The current strategy for doing this was to generate an empty file and
wrap that in an archive.  However, some system linkers (e.g. ld.gold)
require that archives contain valid native objects that generate a valid
archive symbol table.  So, this diff changes this workaround to compile
an empty haskell module and uses that to pack into the dummy archive that
gets passed to the haskell linker.

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: e71eaf3
fbshipit-source-id: e71eaf3