Add tests for annotation processor rebuild behavior
Summary: Most Java deps we can just look at whether the ABI changed and not rebuild if it didn't. Annotation processor deps are different -- they represent code that will be run in order to build the target. Therefore *any* change to any target in the dep graph of an annotation processor must trigger a rebuild of targets that are built with the annotation processor. We had a test hole here; filling it.

Test Plan:
Comment out the line in `AbstractJavacOptions.appendToRuleKey` that adds the annotation processor params to the rule key. Observe
the tests fail.

Reviewed By: Coneko

fb-gh-sync-id: b3e02fa
fbshipit-source-id: b3e02fa