Adding base_module_strip to Python targets
Often times I want to organize my source tree so that my Python sources
start one or two directories down from the project root.  With a
.buckconfig at the top level these two directories must exist in the
module namespace unless you go and manually specify base_module for each
of the targets.  When moving sources into Buck this meant I would need
to refactor a large number of files just for imports.

This adds a new base_module_strip argument to Python targets, which if
specified strips that many elements when PythonUtil.getBasePath is
called using the subpath method on the path.  You can also specify the
value as a config value and have it apply to all targets.