Make JavaDepsFinder operate on TargetNodes, not BuildTargets
Summary: TargetNodes are the actual nodes in a dependency graph; BuildTargets are just ways of addressing them. This makes JavaDepsFinder operate more closely to the actual data, as well as prepares it for TargetNode owning visibility instead of BuildTarget.

Test Plan:
Reference a class not in the dependencies of a target with autodeps = True, verify that Buck is still able to find the target that should be included in deps. Specifically:

1. Add a CommandTool field to ActionGraphCache
2. Run "buck build buck"

Expected results:

Build fails, buck informs user that //com/facebook/buck/rules:action_graph_cache is missing dependency :command_tool

Reviewed By: Coneko

fb-gh-sync-id: 8965e09
fbshipit-source-id: 8965e09