Unexpected flavors handling
Summary:
The error messages you get when the flavor is not recognised are not really useful and easy to understand for the user. A problem that intensifies this is that we do not know what exact flavor caused the issue, just that the given combination is not recognised and supported.

Introducing `UnexpectedFlavorException` gives the ability for better messaging when things go wrong. The plan is that `UnexpectedFlavorException` will hold regex patterns for common flavors and some messages with suggestions on how to fix them. Given that we can't get which flavor caused the issue we will have to give all the suggestions that match our patterns.

The `android-*` and its message is just an example. I expect this to be filled by the people based on the most common needs. Another diff will provide the ability to specify "fix messages" for unrecognised flavors in `.buckconfig` so even better insight can be given.

Test Plan: Added a unit test that tests the suggestion is provided when the flavor's name matches the pattern.

Reviewed By: marcinkosiba

fb-gh-sync-id: a7bcca2
fbshipit-source-id: a7bcca2