Allow disabling scuba logging via buckconfig
Summary:
When used on a slow connection, Buck is overwhelmed by hundreds of Scuba logging attempts that frequently time out. It surprised me to find that in this event, Buck simply drops messages on the floor:

```
[2016-05-07 11:03:40.864][warn ][command:c1f7e722-a949-4738-b7c6-912f6f5a855c][tid:01][com.facebook.buck.event.BuckEventBus] The BuckEventBus failed to shut down within the standard timeout.
Your build might have succeeded, but some messages were probably lost.
```

The end result: Buck finishes builds but does not show the error message that caused the build to fail.

I can consistently reproduce this by throttling my internet connection to high latency and ~100Kbps.

This is rather aggravating. I would attempt to fix this myself, but I'm over my head; I think it might require refactoring the `BuckEventBus` to have a QOS concept so it can drop the least important messages first.

Instead, I am adding an option to disable Scuba logging so I can simply turn it off. Out here on a Wisconsin farm with barely-functioning Internet, I am not that interested in wasting my limited bandwidth submitting hundreds of these to Scuba, so I'll just opt out and let others fill it in!

Test Plan:
Reproduce the problem by running buck build on a slow internet connection. Observe that when there is a compilation error, buck simply exits instead of showing the error, and prints the above notice in buck.log.

Then, turn off scuba logging with this new config switch, and verify that buck build correctly shows the error message when the build fails.

Reviewed By: michath

fbshipit-source-id: 56f5ea0