Remove code for adding extra event listeners.
Summary:
In the past Buck had a feature for defining additional event listeners
in .buckconfig. That feature was removed, but the code remained because
it was used in a few tests. This changes those tests to use other ways
to verify their assertions and removes the last pieces of that code.

Test Plan: CI

Reviewed By: beefon

fbshipit-source-id: 75e620d