Remove defaultRuleKeyBuilderFactory from RuleKeyBuilder
Summary:
Only the `RuleKeyBuilder` implementations in
`DefaultRuleKeyBuilderFactory` use the
`defaultRuleKeyBuilderFactory` field, which the input-based rule
key builders do not use it at all. It does not make sense for them
to have this field and to require an instance in their
constructors. Remove the field from the common superclass and only
put it in the subclasses that actually use it.

Test Plan: `buck test`

Reviewed By: andrewjcg

fbshipit-source-id: de7ee64