Fix dependency resolution when using a WorkerTool
Summary:
This diff fixes two issues related to ##WorkerTool## dependencies:

The first is that a ##WorkerTool## didn't include the rules referenced in its ##args## param as runtime deps, which meant that those dependencies wouldn't be present locally if they were fetched from cache.

The other issue is that a ##Genrule## that referenced a ##WorkerTool## using a ##$(worker)## macro didn't include the ##WorkerTool## itself as a dependency. As a result the dependencies of the ##WorkerTool## (such as the ones referenced in ##args##) weren't guaranteed to be built when building the ##Genrule##.

Test Plan:
Added test cases:
  buck test //test/com/facebook/buck/shell:shell

Reviewed By: Coneko

fbshipit-source-id: 69e08d0