Print stderr again at the end when ShellStep fails
Summary:
When a buck build fails running a ShellStep the stderr form the step often gets hidden by a lot of 'BUILT ...' lines of things that got built on other threads. This will print the stderr again, after the 'BUILD FAILED: ' line.

Changes:
* change the return type of execute in Step interface from int (exit code) to new class StepExecutionResult (that contains exit code and can have stderr added)
* update all classes implementing Step and everything that calls execute to fit the new method signature
* return StepExecutionResult with stderr from ShellStep
* print stderr in StepFailedException if available

Files with non pure refactor code:
* src/com/facebook/buck/step/AbstractStepExecutionResult.java
* src/com/facebook/buck/shell/ShellStep.java
* src/com/facebook/buck/step/StepFailedException.java
* test/com/facebook/buck/cli/BuildKeepGoingIntegrationTest.java

Test Plan:
Run a build that fails on on a ShellStep that prints an error to stderr on fail
Check if the stderr error shows up again after BUILD FAILED line.

And Sandcastle.

Reviewed By: Coneko

fbshipit-source-id: dd224cb