Update .classpath file for Eclipse.
Summary:
This does not quite build on Eclipse Mars Milestone 5,
but it's much closer to building than it was before.
I think this version of Eclipse is too old to pick up the
necessary fixes for the immutables library in 'Upgrade ant build to ecj 4.6M6 and immutables 2.1.28',
as that was declared to be designed to work with 4.6M6.

Test Plan: See summary.

Reviewed By: Coneko

fbshipit-source-id: 8652e9c