Fix incorrect source of class sources being used
Summary:
The orderEntry for the JDK and source code were at the end of the list meaning that any class which could be decompiled to provide the source code for a class was used ahead of the sourceFolder entry which would contain the actual sources.

This diff ensures that when a sourceFolder is present the orderEntry element appears just behind the JDK entry so they are queries for the source before any libraries which may contain compiled versions of the library.

Test Plan:
- Local reference chasing
- Debugging an android application

Reviewed By: marcinkosiba

fbshipit-source-id: bd11f36