Change output type of input rule key builder
Summary:
Make input-based rule key builder return an `Optional<RuleKey>`
so that it has a way to not return any rule key in cases where
it decides it is not worth it (e.g. hashing the inputs would take
longer than building the rule locally).

Test Plan:
New tests to check that the `CachingBuildEngine` handles missing
rule keys properly.

Reviewed By: andrewjcg

fbshipit-source-id: 948c78a