python: package shared libs for root `force_static=True` C/C++ deps
Summary:
We use `force_static=True` as a way to signal to Buck that a C/C++ dep
should be statically linked into any other C/C++ libs which depend
on it.  However, when a python library depends on a C/++ lib with
`force_static=True` set, this hint probably doesn't make sense, and
we should prioritze the explicit dep the user added onto the library.

This allows `force_static=True` to model things like libjemalloc.  When
we're linking a C/C++ binary, it may make sense to always statically
link this into the top-level binary, so that deps' `malloc` calls will
use jemalloc's versions from the main binary.  However, when building
a Python binary, we instead need to packaged libjemalloc as a shared
lib and use `LD_PRELOAD` at startup.

Test Plan: added unittest

Reviewed By: Coneko

fbshipit-source-id: 1a03a81