Record resource ids referenced by libraries we dex
Summary:
In order to trim R.java to only the fields that are actually referenced
from Java, record the fields referenced by each each library we pre-dex.
This will only work in buidls with pre-dexing on, but that's okay for
now.  This is the fastest way to get this list because we have all of
the referenced fields for an entire module in one easy-to-digest list.

Test Plan: Integration test.

Reviewed By: marcinkosiba

fbshipit-source-id: 250f180