Switch network executor to same WorkQueue mode as main thread pool.
Summary:
Network executor was not using the same WorkQueue as the main thread pool we
have. This introduced the problem where the main thread pool is usually on LIFO
mode and the network executor was on FIFO. The result was a regression on
performance of the CachingBuildEngine.

Test Plan: * CI

Reviewed By: Coneko

fbshipit-source-id: 376b0ba