Add supported_platforms_regex to PrebuiltCxxLibrary
Summary:
The new parameter to prebuilt_cxx_library is needed for the exact same use case as cxx_library. We add it, along with the relevant documentation and a unit test (which is very similar to cxx_library's version).

This is necessary for supporting compiling for armv5 - some libs do not support it, and we shouldn't be trying to include them.

Test Plan: Unit test. Also was able to solve the first problem explained here: https://fb.facebook.com/groups/askbuck/permalink/1069006206481327/?comment_id=1072106639504617

Reviewed By: Coneko

fbshipit-source-id: b192de5