De-flake CachingBuildEngineTest
Summary:
The assertRelatedBuildRuleEventsOnSameThread method in the test assumed
that only one thread executes operations related to a single rule at any
given time. This is not true: the scope in processBuildRule can be active
at the same time as ruleAsyncFunction is running (this happens if the
thread created for ruleAsyncFunction preempts the processBuildRule thread).
Rather than asserting properties of the total ordering of events related to
a single BuildRule we will assert properties of events submitted to a single
thread.

Test Plan: run flakey test 5000 times

Reviewed By: michath

fbshipit-source-id: cb26ec1