Add a per-cell DaemonicState
Summary:
When multiple cells are in play, the target parsing invalidates the
cache more than it should, causing some keys in the cache to go
missing. This is because a change in one cell spills over to the other
cells when invalidating. The real thing we want is to invalidate the
DaemonicParserState based on cell.

Test Plan: Unit tests still pass.  Add a new one

Reviewed By: marcinkosiba

fbshipit-source-id: dd5bb6d