Add ability to ban duplicate string resources in app
Summary: We want to ban duplicate string resources in our Android app because this is a common cause of subtle bugs where the wrong string is used. There are no intentional use cases for duplicate strings anymore because it is brittle (dependent on order of buck dependencies to work). However, it is very easy to add duplicate string resources so we have a lot of problems to fix. We're adding this as a configuration that can be enabled at the app level so we can roll this out one app at a time.

Test Plan: Tests, build an app with duplicate string resources disabled and see that the build breaks.

fbshipit-source-id: 5fe69c3