Introduction of remove_classes to java_library & usage on directSpoolToJar
Summary:
This enables the ability to remove Class files from the Jar for java_library. Now java_library has another field called `remove_classes` that is a list of patterns. If one of the patterns matches the source file path it is removed from the source path list and never gets into the JAR.

+ Documentation

Test Plan:
* CI
* New unit test that removes the member classes of 2 classes from the Jar.

{F61310340}

Reviewed By: marcinkosiba

fbshipit-source-id: 862747d