Include worker_tool dependencies when computing InputBasedRuleKeys
Summary:
If a ##genrule## referenced a ##worker_tool## with a ##$(worker)## macro, the ##genrule## might not always rebuild if a dependency of the ##worker_tool## changed. This is because ##Genrule## implements ##SupportsInputBasedRuleKey## and the ##Tool## returned by ##WorkerMacroExpander.extractRuleKeyAppendables## wasn't including the ##worker_tool##'s dependencies as inputs. This diff corrects that.

In order to avoid a dependency cycle, ##WorkerTool## is now an interface with ##DefaultWorkerTool## as the concrete implementation.

Test Plan:
Added a unit test:
  buck test //test/com/facebook/buck/shell:shell

Reviewed By: Coneko

fbshipit-source-id: 14a580e