Clean up exception handling for WorkerProcess closing.
Summary:
Turns out that destroying the process can also throw (which is especially
likely if there were previous issues), so we end up with a
HumanReadableException that is suppressed by the IOException from the finally
block.
To address this we move the try/catch one level higher.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 468b372