Add archive member hashes to distributed build state.
Summary:
This makes the archive member hashes part of the distributed build state.
The alternative was to have the remote host download the archive and hash
its members, however that would mean the remote would need to do so very
early in the build and all other operations would be blocked by the fetch.

Test Plan: added new test

Reviewed By: ruibm

fbshipit-source-id: eff5911