Move pattern matching logic to a helper class
Summary:
Extracting logic responsible for matching attributes to a set of patterns into a separate helper class. The primary reason to do so is to reuse this class in `TargetsCommand` to provide the same feature (filtering output attributes in JSON format).

One nice bonus of this refactoring is compiling each pattern only once, not every time for every target.

Test Plan: Unit tests and local runs.

fbshipit-source-id: 1fe8a0c