omnibus: include excluded root deps as omnibus deps
Summary:
The way which we make undefined root symbols affect the omnibus link
means that these symbols get emitted as undefined symbols in the omnibus
library.  This means we need to also need to propagate all root deps to
the omnibus library, which we weren't currently doing.

Test Plan: updated unittests

Reviewed By: yiding

fbshipit-source-id: b6e7d89